project_root/requirements.txt
Flask
python-dotenv
requests
Pillow
pytesseract
opencv-python-headless
pytest
google-genai
scikit-learn
sentence-transformers
numpy
Pillow
python-dotenv  # Recommended for handling API keys securely